Move AnnotatedInstanceSetHistory
