+++ Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test/Google.Solutions.IapDesktop.Extensions.LogAnalysis.Test.csproj
@@ -108,10 +108,10 @@
     <Compile Include="FixtureBase.cs" />
     <Compile Include="History\TestDataPoint.cs" />
     <Compile Include="History\TestInstanceHistoryBuilder.cs" />
-    <Compile Include="History\TestAnnotatedInstanceSetHistory.cs" />
+    <Compile Include="Services\UsageReport\TestAnnotatedInstanceSetHistory.cs" />
     <Compile Include="History\TestInstanceSetHistoryBuilder.cs" />
     <Compile Include="History\TestInstancePlacement.cs" />
-    <Compile Include="History\TestLicenseLoader.cs" />
+    <Compile Include="Services\UsageReport\TestLicenseLoader.cs" />
     <Compile Include="History\TestNodeHistory.cs" />
     <Compile Include="History\TestNodeSetHistory.cs" />
     <Compile Include="History\TestTimeseriesUtil.cs" />
