+++ Google.Solutions.IapDesktop.Extensions.LogAnalysis/Services/UsageReport/ImageAnnotation.cs
@@ -24,7 +24,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using System;
 
-namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.History
+namespace Google.Solutions.IapDesktop.Extensions.LogAnalysis.Services.UsageReport
 {
     [Flags]
     public enum LicenseTypes
